Fixed incorrect use of PICK-ONE
The PICK-ONE routine takes an LTABLE with an extra element at the start
that is initialized to 0. This is used for bookkeeping to ensure that no
element is repeaed until every element in the table has been used once.

Most noticeably, the game no longer crashes when exiting the Carousel
Room. The other two incorrect uses were the random hallucinations caused
by the Fantasize spell and the description for the Wizard's Quarters.