Set the balloon location correctly when landing on a ledge.
Before, this was alwas set to 0 which meant that taking off from a ledge
would move the balloon to the Volcano Core.

When "walking" inside the balloon (and note that "LAND" is a walking
direction here), it checks if the current room has an unconditional exit
in that direction. If so, the balloon location is set to that exit's
destination.

This is what the full game does, unless the destination has RMUNGBIT
(i.e. the room has been destroyed by an explosion or similar). This
version of the game doesn't use RMUNGBIT. I'm guessing the assignment
was accidentally lost when that piece of code was removed.