Don't automatically take the fuse on "LIGHT FUSE"
This bug happened because the "LIGHT OBJECT" syntax specified both TAKE
and HAVE for the object to light. Since the verb function is V-LAMP-ON,
I assume it's primarily intended for the lamp, but at some point someone
decided - correctly, I think - that "LIGHT FUSE" is a reasonable
command.

What's not so reasonable is that the parser then automatically took the
fuse out of the brick, leaving you holding the burning fuse with not
enough time to put it back into the brick and make it out alive. This
feels like a Wile E. Coyote moment to me.

By giving the fuse TAKEBIT, the parser no longer tries to automatically
take it, and by leaving out the HAVE flag on the syntax you are no
longer required to hold it.

I can't think of anything that would break from this change. Worst case
would be that you manage to light the lamp without carrying it. But you
can already do that with "TURN ON LAMP". (Perhaps because "TURN ON ..."
could sensibly refer to some big immovable piece of equipment?)