Kill the player when burning a held object.
This was obviously the intention (a very cheap and Zorkian death!), and
it's what the full game does. Note that this refers to the default
behavior. Objects can still override it, e.g. you can still safely burn
the fuse while holding it.

I guess the call to REMOVE-CAREFULLY was moved so it would only have to
be called once (rather than once if the object was held and once if it
wasn't), without considering that this would screw up the check for if
the object was held.

With this fix, the burned object will first be moved (by JIGS-UP) and
then removed, but that should not be a problem.