Some fixes to the key/mat puzzle.
Mention if there is a key in the keyhole when looking through the red
sphere. Unlike the full game, the keyholes have no capacity so there
shouldn't ever be any other objects than the key in them.

When the key falls on the mat, remove the key. Otherwise it's listed
twice if you look into the sphere again.