Fixed logic error in I-TOPIARY.
I-TOPIARY only checked if TOPIARY-COUNTER was 0. That meant that you
would only ever get one of the topiary messages, and that it might kill
you straight away instead of warning you first.

In the full game, there are two global variables (TOPIARY-MOVED and
TOPIARY-NEAR), so the error must have snuck in when this was simplified
to use just one global variable.