Only let the Wizard see the demon if he's visible.
Normally, the Wizard stops appearing after he sees you talking to the
demon. But if you ask the demon for the wrong favor (e.g. moving the
menhir) the demon disappears and the Wizard remains.

If you then attack the Wizard, the I-WIZARD daemon is restarted. Since
the demon is still in the pentagram room, it was possible for the wizard
to see him a second time.