Various fixes for the menhir.
In the Menhir Room, always describe the menhir even if it's no longer in
LOCAL-GLOBALS. That's because the description will still refer to the
passage, even if the menhir has been moved out of the room (by asking
the demon to get it or give it to you). Oddly enough, the menhir was
already always described in brief mode.

Added "large" as adjective for the menhir objects, since it can be
described as a "large menhir".

The menhir was described incorrectly when floating in the air. The
description depends on MENHIR-POSITION:

0 - The initiali state.
1 - The demon had moved the menhir, but it's still in the Menhir Room.
2 - The demon had taken the menhir, removing it from the game.
3 - The demon had given the menhir to you, moving it to the Pentagram
    Room, or you had used the wand to levitate the menhir.

But the menhir was only described as floating if MENHIR-POSITION was
none of these. Since there's no difference between the menhir being gone
because the demon took it or because it had been moved to the Pentagram
Room (at least not as far as the Menhir Room is concerned), I have
changed it so that both these conditions use MENHIR-POSITION 2.

Finally, I've changed the name of the menhir objects from "enormous
menhir" to "huge menhir". Partly to preserve the now removed line "There
is a huge menhir here." and partly to avoid the game describing it as "a
enormous" [sic] menhir. Of course, that could be fixed but it seems
unnecessary for a game that should be kept small.

If memory serves me, one or more of these issues are present in the full
game as well.