Fixed the menhir being visible even after being removed.
It turns out that putting local global object inside LOCAL-GLOBALS is
just convention, not a rule. If you ask the demon to give you the
menhir, he'll move it to the Pentagram room but it's still visible from
any room where it's local global.

There are several ways this could be fixed: My first idea was to change
the parser, but I'd like to stay away from that. Changing MENHIR-F seems
like a less intrusive change.

So now the menhir is only "visible" if it's in the room with you, or if
it's detected as a local global object.