Rebalanced the score slightly to get the total to 400 points.
The game only awarded 392 points. The following changes have been made:

* Shrinking in the Alice area is worth 5 points (was 10)
* Reaching the Crypt is 5 points (was 2)
* Reaching the Volcano Core is worth 10 points (was 0)

Some notes about differences in scoring between this and the full game:

Treasures and other valuables are scored the exact same way, without any
regard for some of them being a lot easier to find.

The full game awards 5 points for solving the maze, and another 10 for
reaching Cerberus' room. Since there is no maze in this version,
reaching Cerberus' room isn't much of a puzzle so no points are awarded
for that.

Instead you get points for shinking yourself in the Alice area, and for
getting the balloon to work.

The Crypt Anteroom and Crypt are combined into one Crypt room, so the
points for reaching them are combined into one.

One other interesting difference is that in the full game, getting the
Wizard's wand is likely to put you over 360 points, earning you the
second highest rank of "Wizard". I thought that was a neat touch.

In this version, the number of ranks has been reduced, and "Wizard" is
the higest rank. You won't see that until the very end of the game.