Hopefully fixed the Wizard's Fantasize spell.
It appears that this version of the game has the same Fantasize bug as
the final preserved version of the Zork II source code has, in that the
hallucination is only printed if there are objects in a container in the
room.

Fortunately, an older version was preserved that doesn't exhibit this
bug. I'm guessing it was introduced by accident when a common
"substrate" was developed for the Zork trilogy. (In the final release, I
think the code for handling the Fantasize spell had been acidentally
dropped completely.)

These changes are based on changes I've proposed for the full game. I
don't fully understand the technical details, but hallucinations seem to
appear when they should, and some strange indentation for containers has
been fixed based on how it's done in later games.