Removed code that dealt with different types of vehicles.
It was possible to define several types of vehicles by specifying a bit
in the vehicle's VTYPE property. But both vehicles (the bucket and the
balloon) in this game always used the same bit, NONLANDBIT, for this.

The simplified rule is as follows:

If you are on land (RLANDBIT) in a vehicle, you can only travel to
NONLANDBIT rooms. (This works out even if you are killed, because
JIGS-UP moves you out of the vehicle first.)

The only rooms in the game that don't have RLANDBIT are VOLCANO-CORE,
VOLCANO-BY-SMALL-LEDGE and VOLCANO-BY-WIDE-LEDGE. In these rooms, you
have to be able to travel freely because you land by walking in a
direction.

All rooms you can travel to by vehicles have NONLANDBIT, including the
top and bottom of the well.

Also, there is no longer a message when the vehicle comes to a stop,
because that's already handled elsewhere. (This fixes double messages
when landing the balloon.)