Generalized doors a bit.
Made OPEN-CLOSE take an argument so that the caller doesn't have to
print "The door is locked."

Similarly, added UNLOCK-LOCK to handle unlocking and locking doors. This
change isn't completely invisible to the user, but I think it makes
sense to have them use the same messages. Particularly when the
difference was as small as it was before, and the messages were not in
any way unique or interesting.