Fixed the game printing garbage on "DRINK WATER".
"DRINK WATER" would always refer to WATER, not GLOBAL-WATER, by the time
the action was handled. There it checks if the player is carrying the
container the water is inside. If the WATER object wasn't in any
containing object, it would print garbage.

Now the EAT and DRINK actions bypass most of the WATER-F logic.

I've also added WATER as a synonym to the pool. It's a pool of tears in
the full game, where you can use "pool" or "tears". It's a pool of water
in this version.