Fixed the behavior of "EXAMINE ALL" and similar.
For whatever reason, some actions were defined as "OBJECT (MANY)" which
apparently allows them to work on far too many objects. I've made some
changes to correspond with the full game.

For most of these, it was just a matter of changing "OBJECT (MANY)" to
"OBJECT (HELD CARRIED IN-ROOM ON-GROUND MANY)". In the case of "LOOK
WITH OBJECT" the original has that one as just "OBJECT", so I removed
the "(MANY)" completely there.