Simplified V-LEAP. There are no CEXITs leading down in this game.
The V-LEAP routin checked if there it a DOWN exit in the current
location that is either an NEXIT or a CEXIT.

An NEXIT simply prints a string, e.g. (DOWN "It's a long way down.").

A CEXIT checks a condition, e.g. (DOWN TO WELL-BOTTOM IF TRAMPOLINE-FLAG
ELSE "That would be suicidal.")

Both of these are taken to mean that there is an exit leading down, but
it would be fatal to use. Jumping in such a room kills you. But, as
already said, this game has no downard CEXITs. Nor does the full game,
as far as I can tell.