Removed PRE-MUNG.
The game often (though not always) treasure MUNG and ATTACK as the same
thing, so it seems strange that it would let you attack without a
weapon, but not destroy without one.

This prevented GLOBAL-PALANTIR-F from responding to the MUNG action.
Though in order for it to work, I also had to add a syntax for "DESTROY
OBJECT" (i.e. without an indirect object).

The messages from PRE-MUNG now reside in V-MUNG.

I hope this doesn't cause any regressions.