Removed special case fro ME-F.
It looks like at some point, "GIVE object TO ME" was supposed to be a
synonym for "TAKE object". I've seen it in other games, but I don't
think I've ever seen it work. (The syntax for GIVE requires you to have
the object already.)

I don't think this behavior is interesting enough to even try to
restore. Let's save the space instead.