Don't print "The match has gone out." unless you can see it.
It is possible to get rid of the matchbook (e.g. dropping it down a
chasm) before it burns out. Also, inform the player if the match was
your last source of light.

In addition, JIGS-UP used to disable the I-MATCH timer, probably so you
wouldn't see the match go out if you just died. However, this would turn
the matchbook into an infinite light source, so let's fix that.