Various fixes to the balloon receptacle.
Much of the special cases for the balloon receptacle were handled in
BALLOON-F, when RARG was M-BEG. But this is only invoked when you are
actually inside the balloon. If you were doing it from outside the
basket, they were not triggered.

The most noticeable effect was that any object put inside the receptacle
from outside the balloon did not get the NDESCBIT, so it was described
both in the balloon description and listed as part of the objects in the
room.

I have fixed this by adding an action routine and a container routine
for the receptacle. As an addition bugfix, now when you look inside the
receptacle the content's NDESCBIT is temporarily cleared. It's not the
most exciting of messages, but at least it will no longer be described
as empty.