Show the ballon's M-LOOK description after landing on the ledges.
DESCRIBE-ROOM would return after calling a room's M-LOOK action routine
(if it returned true), before the vehicle's M-LOOK action routine could
be called. This may be by design, but it meant that the ballon
description wasn't printed after landing on either of the volcano
ledges.

In the full game, this only happens on one of the ledges. The other uses
LDESC instead of an action routine. But in this version, it happened on
both ledges so I think it warrants a fix.