Removed the last traces of "FOLLOW PRINCESS".
In the full game, there is a global princess object that allows you to
"FOLLOW PRINCESS" after she leaves the room. But since that object has
been removed, along the code that makes sure you can't do anything else
with the princess when she isn't there, it doesn't work.

Rather than putting back the code that was clearly removed for space,
disable the remaining bits dealing with following the princess.

Other Infocom games handle following without having a global object like
this, but the infrastructure for that doesn't seem to be present in this
game, so that's not an option either.