Don't show Cerberus as floating in mid-air.
If you levitate Cerberus, he just rises an inch of the ground for a
moment. The game sets SPELL-HANDLED? to suppress the DISENCHANT action
when the spell times out, but Cerberus is still the SPELL-VICTIM.

To keep the wand recharge timeout, but not show Cerberus as levitating,
SPELL-VICTIM is now set to a different object. GLOBAL-CERBERUS seemed
the most fitting.