Preserve TRYTAKEBIT across growing/shrinking in the Alice area.
When you shrink in the Alice area, all objects in the room are given
TRYTAKEBIT so that you can't pick them up in any way. When you grow to
normal size, TRYTAKEBIT was indiscriminately cleared. That meant that
objects that started out with TRYTAKEBIT no longer had it.

Now it uses RLANDBIT to store the old value of TRYTAKEBIT, so that it
can clear it selectively.