Changed PRE-TAKE special case for DOOR-KEEPER based on the full game.
In the latest version of the full game, PRE-TAKE checks if PRSI is the
GROUND or if PRSO is the DOOR-KEEPER. Not if PRSI is the DOOR-KEEPER.

My only guess is that this is to deal with "TAKE LIZARD FROM DOOR", or
something like that. It makes marginally more sense than dealing with
"TAKE object FROM LIZARD". Anyway, it's a small change to keep the games
consistent with each other.