Fixed double message in I-WIZARD.
If the Wizard appears in a dark room, there already is a message about
movement in the dark. There's no need to print a second identical
message if he retreats because you're carrying the black sphere.

In the full game, these are two different messages so there it makes
sense to have them both.