Hid the DISENCHANT action from the player.
When a spell times out, it generatet a DISENCHANT action. This was a
neat way to make objects have their own reaction, i.e. there's a custom
message for the menhir floating to the ground.

But that meant the player could type "DISENCHANT object", and this did
not work because part of the actual disenchanting is in I-SPELL.

So I've changed the old DISENCHANT action to $DISENCHANT, to mark it as
an internal command. The player can still type it, but it's unlikely to
happen by mistake.

A new DISENCHANT action has been added, because it's still a sensible
thing to type.