Reinstated handling of "PUT SERPENT IN object".
The full game's message, "Impossible for many reasons." makes more sense
this way. Note that this required a slight change to the syntax for PUT,
so that it doesn't require the objects to be held. The PRE-PUT routine
should check that anyway.

Other actions that map to the PUT action, e.g. "THROW OBJECT IN ...",
"DROP OBJECT DOWN ...", etc. remain unchanged.