Implemented VOWELBIT.
This is a bit wasteful, but it looks odd to me to carry "A elvish
sword". Later Infocom games used VOWELBIT to fix this. I hope I caught
all the cases where it should be used.

This also means that the menhir objects can be "enormous menhir" again,
as they originally were.