Made RANDOM-BOOK-F handle the OPEN action as well.
Since opening an object sets its TOUCHBIT, opening the purple book would
cause it to be listed after the white book, even though the white book
still referred to the purple book in its description. Now all three
books get TOUCHBIT, just like for the TAKE, MOVE and PUT actions.

(Actually, I'm not sure what the point of the PUT action is here, but
let's keep it just to be safe.)