Restore the old VERBOSE / SUPER-BRIEF variables.
Whoever made Mini-Zork II had introduced a new variable, VERBOSITY, to
replace the old VERBOSE / SUPER-BRIEF variables. On paper this seems
sensible, but it actually uses more space that way.

Reverting the change also brings the parser a bit closer to the final
parser for the three full Zork games, which is probably where it
originated.