Moved "with <object> is <something>" into its own routine.
This saves a tiny bit. If I tried to move the whole "Trying to ..."
message into a single routine, it did now. Possibly because I had to
distinguish between if PRSO is printed as "a object" or "the object".

It's quite possible PRSO would have been printed the same way in both
actions if work had continued at Infocom, but I'm not going to make that
change here.